+++ precompiles/oracle/oracle.go
@@ -92,6 +92,15 @@ func (p Precompile) RequiredGas(input []byte) uint64 {
 
 // Run executes the oracle precompile
 func (p Precompile) Run(evm *vm.EVM, contract *vm.Contract, readOnly bool) (bz []byte, err error) {
+	bz, err = p.run(evm, contract, readOnly)
+	if err != nil {
+		return cmn.ReturnRevertError(evm, err)
+	}
+	return bz, nil
+}
+
+// run executes the oracle precompile (internal implementation)
+func (p Precompile) run(evm *vm.EVM, contract *vm.Contract, readOnly bool) (bz []byte, err error) {
 	// Initialize the context, db and chain data
 	ctx, _, method, initialGas, args, err := p.RunSetup(evm, contract, readOnly, p.IsTransaction)
 	if err != nil {