+++ x/feeabstraction/ante/cosmos/fee.go
@@ -140,6 +140,9 @@ func (dfd DeductFeeDecorator) checkDeductFee(ctx sdk.Context, sdkTx sdk.Tx, fee 
 			return err
 		}
 
+		// Refresh info, it can't be nil since we checked it exists before
+		deductFeesFromAcc = dfd.accountKeeper.GetAccount(ctx, deductFeesFrom)
+
 		// Deduct the fees from the fee payer account
 		err = ante.DeductFees(dfd.bankKeeper, ctx, deductFeesFromAcc, convertedFee)
 		if err != nil {