+++ x/feeabstraction/keeper/abci.go
@@ -44,14 +44,19 @@ func (k Keeper) WriteFeeTokenPricesMetrics(ctx context.Context) error {
 	}
 
 	// Iterate over the fee token prices and set the gauge metrics
-	for _, price := range feeTokenPrices.Items {
-		// Set a module metric for each token
-		if floatPrice, err := price.Price.Float64(); err == nil {
+	for _, token := range feeTokenPrices.Items {
+		// If token is disabled, skip to next one
+		if !token.Enabled {
+			continue
+		}
+
+		// Set a module metric for each enabled token
+		if floatPrice, err := token.Price.Float64(); err == nil {
 			telemetry.ModuleSetGauge(
 				types.ModuleName,
 				float32(floatPrice),
 				"fee_token_price",
-				price.Denom,
+				token.Denom,
 			)
 		}
 	}