+++ x/feeabstraction/types/tx.pb.go
@@ -366,6 +366,7 @@ func _Msg_UpdateFeeTokens_Handler(srv interface{}, ctx context.Context, dec func
 	return interceptor(ctx, in, info, handler)
 }
 
+var Msg_serviceDesc = _Msg_serviceDesc
 var _Msg_serviceDesc = grpc.ServiceDesc{
 	ServiceName: "kiichain.feeabstraction.v1beta1.Msg",
 	HandlerType: (*MsgServer)(nil),