+++ CHANGELOG.md
@@ -5,10 +5,14 @@
 ## Added
 
 - Cosmos EVM integrations tests added to the repo
+- Added logs and telemetry for reentrance detection in wasmd precompile
+- Add further validations to Wasm Oracle query bindings
+- Moves Wasmd reentrance lock to the core of the wasmd contract to avoid reentrance attacks on queries and instantiations
 
 ## Removed
 
 - Removed IBC precompile since ICS20 precompile also handles IBC transfers.
+- Removed fallback native price param on the fee abstraction module. Instead of using hardcoded price, the module disables itself when price is lacking.
 
 ## DEPENDENCIES
 
@@ -22,6 +26,7 @@
 - Fix IBC validation of negative numbers happens a bit early [#144](https://github.com/KiiChain/kiichain/issues/144)
 - Fix incorrect error passing on tokenfactory wasmbinding
 - Fix blocked address checked after minting in wasmbinding tokenfactory 
+- Fix account balance information being stale on fee abstraction's cosmos' ante
 
 ### Documentation